+++ scripts/cassandra/ICassandraClient.ts
@@ -2,7 +2,7 @@ import {Observable} from "rx";
 
 interface ICassandraClient {
     execute(query:string):Observable<any>;
-    stream(query:string):Observable<any>;
+    paginate(query:string, completions:Observable<void>):Observable<any>;
 }
 
 export default ICassandraClient