+++ scripts/streams/EventsFilter.ts
@@ -4,6 +4,7 @@ import * as _ from "lodash";
 import IEventsFilter from "./IEventsFilter";
 import {Matcher} from "../matcher/Matcher";
 import {helpers} from "rx";
+import ReservedEvents from "./ReservedEvents";
 
 @injectable()
 class EventsFilter implements IEventsFilter {
@@ -11,12 +12,17 @@ class EventsFilter implements IEventsFilter {
     private events:string[] = [];
 
     filter(definition:IWhen<any>):string[] {
-        if (definition.$any) return this.events;
-        let matcher = new Matcher(definition);
-        let list = _(this.events).map(event => {
-            return matcher.match(event) !== helpers.identity ? event : null;
-        }).compact().valueOf();
-        return !list.length ? this.events : list;
+        let eventsList:string[];
+        if (definition.$any) {
+            eventsList =  this.events;
+        } else {
+            let matcher = new Matcher(definition);
+            eventsList = _(this.events).map(event => {
+                return matcher.match(event) !== helpers.identity ? event : null;
+            }).compact().valueOf();
+            eventsList = !eventsList.length ? this.events : eventsList;
+        }
+        return _.union(eventsList, [ReservedEvents.TICK, ReservedEvents.REALTIME, ReservedEvents.FETCH_EVENTS]);
     }
 
     setEventsList(events:string[]) {