+++ scripts/streams/ReservedEvents.ts
@@ -1,4 +1,5 @@
 export default {
     TICK: "Tick",
-    REALTIME: "__prettygoat_internal_realtime"
+    REALTIME: "__prettygoat_internal_realtime",
+    FETCH_EVENTS: "__prettygoat_internal_fetch_events"
 }